+++ bff/src/modules/apiClients/users_game_stats.ts
@@ -1,18 +1,11 @@
 import axios from "axios";
-
-import { isUUID } from "../common/validators/uuid.js";
 import { configService } from "../../envConfig.js";
 import { handleError } from "../common/helpers/handleError.js";
 import type { DeleteUsersGameStatsPayload } from "../models/dto/payloads/users_game_stats/index.js";
 import type { CreateUsersGameStatsResponse } from "../models/dto/responses/users_game_stats/index.js";
 import type { NormalizedError } from "../models/dto/errors/index.js";
-import {
-  ConflictError,
-  AuthError,
-  ValidationError,
-  NotFoundError,
-  ServiceUnavailableError,
-} from "../common/errors/index.js";
+import { isUUID } from "../common/validators/uuid.js";
+import { ValidationError } from "../common/errors/validation.js";
 
 export class UsersGameStatsClient {
   private createEndpoint = "/users-game-stats";
@@ -24,9 +17,7 @@ export class UsersGameStatsClient {
     this.CRUD_API = configService.CRUD_API || "error";
   }
 
-  public async createStats(
-    token: string
-  ): Promise<CreateUsersGameStatsResponse | NormalizedError> {
+  public async createStats(token: string): Promise<CreateUsersGameStatsResponse | NormalizedError> {
     try {
       const response = await axios.post<CreateUsersGameStatsResponse>(
         `${this.CRUD_API}${this.createEndpoint}`,
@@ -35,7 +26,7 @@ export class UsersGameStatsClient {
           headers: {
             Authorization: `Bearer ${token}`,
           },
-        }
+        },
       );
       return response.data;
     } catch (error) {
@@ -45,8 +36,11 @@ export class UsersGameStatsClient {
 
   public async deleteStats(
     token: string,
-    payload: DeleteUsersGameStatsPayload
+    payload: DeleteUsersGameStatsPayload,
   ): Promise<NormalizedError | void> {
+
+    if(!isUUID(payload.statsId)) throw new ValidationError('Invalid UUID')
+
     try {
       await axios.delete(`${this.CRUD_API}${this.deleteEndpoint}`, {
         headers: {
